utilização do estado global
