+++ pokedex/src/components/global/GlobalContext.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const GlobalContext = React.createContext();
+
+export default GlobalContext;
