+++ pokedex/src/routes/Router.js
@@ -21,7 +21,7 @@ const Router = () => {
                     <Route exact path={'/pokedex'}>
                         <Pokedex />
                     </Route>
-                    <Route exact path={'/details'}>
+                    <Route exact path={'/details/:id'}>
                         <DetailsPage />
                     </Route>
                     <Route >
